+++ src/PhpWord/Element/Bookmark.php
@@ -45,9 +45,7 @@ class Bookmark extends AbstractElement
      */
     public function __construct($name)
     {
-
         $this->name = CommonText::toUTF8($name);
-        return $this;
     }
 
     /**